+++ R/findNextCATItem.R
@@ -38,9 +38,9 @@ print.mirtCAT_design <- function(x, ...){
 findNextCATItem <- function(person, test, design){
     
     #heavy lifty CAT stuff just to find new item
-    criteria <- design$criteria
+    criteria <- design@criteria
     if(all(is.na(person$responses)))
-        return(design$start_item)
+        return(design@start_item)
     lastitem <- sum(!is.na(person$items_answered))
     not_answered <- is.na(person$responses)
     which_not_answered <- which(not_answered)
@@ -57,13 +57,13 @@ findNextCATItem <- function(person, test, design){
             row <- row + 1L   
         }
     }
-    method <- design$criteria_estimator
+    method <- design@criteria_estimator
     #saftey features
     if(length(unique(na.omit(person$responses))) < 2L) method <- 'MAP'
     if(sum(!is.na(person$responses)) < 5L) method <- 'MAP'
-    if(design$use_content){
-        tmp <- table(design$content[!is.na(person$responses)])
-        design$content_prop_empirical <- as.numeric(tmp/sum(tmp))
+    if(design@use_content){
+        tmp <- table(design@content[!is.na(person$responses)])
+        design@content_prop_empirical <- as.numeric(tmp/sum(tmp))
     }
     
     if(criteria == 'seq'){
@@ -71,11 +71,11 @@ findNextCATItem <- function(person, test, design){
     } else if(criteria == 'random'){
         if(length(which_not_answered) == 1L) item <- which_not_answered
         else item <- sample(which_not_answered, 1L)
-        if(design$use_content){
-            dif <- design$content_prop - design$content_prop_empirical
+        if(design@use_content){
+            dif <- design@content_prop - design@content_prop_empirical
             tmp <- names(dif)[max(dif) == dif]
             if(length(tmp) > 1L) tmp <- tmp[sample(1L:length(tmp), 1L)]
-            cpick <- design$content[which_not_answered]
+            cpick <- design@content[which_not_answered]
             if(sum(cpick == tmp) > 1L)
                 item <- sample(which_not_answered[cpick == tmp], 1L)
             if(sum(cpick == tmp) == 1L)
@@ -85,31 +85,31 @@ findNextCATItem <- function(person, test, design){
         return(as.integer(item))
     } else if(criteria == 'KL'){
         crit <- KL(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
-                   person=person, test=test, row_loc=row_loc, delta=design$KL_delta)
+                   person=person, test=test, row_loc=row_loc, delta=design@KL_delta)
         index <- which_not_answered
     } else if(criteria == 'KLn'){
             crit <- KL(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
                        person=person, test=test, row_loc=row_loc, 
-                       delta=design$KL_delta*sqrt(sum(!is.na(person$responses))))
+                       delta=design@KL_delta*sqrt(sum(!is.na(person$responses))))
             index <- which_not_answered
     } else if(criteria == 'IKL'){
         crit <- IKL(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
-                   person=person, test=test, row_loc=row_loc, delta=design$KL_delta)
+                   person=person, test=test, row_loc=row_loc, delta=design@KL_delta)
         index <- which_not_answered
     } else if(criteria == 'IKLP'){
             crit <- IKL(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
-                        person=person, test=test, row_loc=row_loc, delta=design$KL_delta,
+                        person=person, test=test, row_loc=row_loc, delta=design@KL_delta,
                         den=TRUE)
             index <- which_not_answered
     } else if(criteria == 'IKLn'){
         crit <- IKL(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
                    person=person, test=test, row_loc=row_loc, 
-                   delta=design$KL_delta*sqrt(sum(!is.na(person$responses))))
+                   delta=design@KL_delta*sqrt(sum(!is.na(person$responses))))
         index <- which_not_answered
     } else if(criteria == 'IKLPn'){
         crit <- IKL(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
                     person=person, test=test, row_loc=row_loc, 
-                    delta=design$KL_delta*sqrt(sum(!is.na(person$responses))))
+                    delta=design@KL_delta*sqrt(sum(!is.na(person$responses))))
         index <- which_not_answered
     } else if(criteria == 'MI'){
         crit <- MI(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
@@ -132,7 +132,7 @@ findNextCATItem <- function(person, test, design){
                      person=person, test=test, row_loc=row_loc)
         index <- which_not_answered
     } else if(criteria == 'Drule' || criteria == 'DPrule'){
-        if(design$numerical_info){
+        if(design@numerical_info){
             crit <- -Drule2(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
                           person=person, test=test, row_loc=row_loc, method=method)
             index <- row_loc
@@ -142,7 +142,7 @@ findNextCATItem <- function(person, test, design){
             index <- which_not_answered
         }
     } else if(criteria == 'Erule' || criteria == 'EPrule'){
-        if(design$numerical_info){
+        if(design@numerical_info){
             crit <- -Erule2(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
                           person=person, test=test, row_loc=row_loc, method=method)
             index <- row_loc
@@ -152,7 +152,7 @@ findNextCATItem <- function(person, test, design){
             index <- which_not_answered
         }
     } else if(criteria == 'Trule' || criteria == 'TPrule'){
-        if(design$numerical_info){
+        if(design@numerical_info){
             crit <- Trule2(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
                           person=person, test=test, row_loc=row_loc, method=method, design=design)
             index <- row_loc
@@ -166,7 +166,7 @@ findNextCATItem <- function(person, test, design){
                       person=person, test=test, row_loc=row_loc, method=method, design=design)
         index <- which_not_answered
     } else if(criteria == 'Wrule' || criteria == 'WPrule'){
-        if(design$numerical_info){
+        if(design@numerical_info){
             crit <- Wrule2(which_not_answered=which_not_answered, possible_patterns=possible_patterns,
                           person=person, test=test, row_loc=row_loc, method=method, design=design)
             index <- row_loc
@@ -179,17 +179,17 @@ findNextCATItem <- function(person, test, design){
         stop('Selection criteria does not exist')
     }
     
-    exposure <- design$exposure[lastitem+1L]    
-    if(design$use_content){
-        dif <- design$content_prop - design$content_prop_empirical
+    exposure <- design@exposure[lastitem+1L]    
+    if(design@use_content){
+        dif <- design@content_prop - design@content_prop_empirical
         tmp <- names(dif)[max(dif) == dif]
         if(length(tmp) > 1L) tmp <- tmp[sample(1L:length(tmp), 1L)]
-        cpick <- design$content[which_not_answered]
+        cpick <- design@content[which_not_answered]
         pick <- cpick == tmp
         if(sum(pick) > 0L){            
             index <- index[pick]
             crit <- crit[pick]
-            exposure <- min(design$exposure[lastitem+1L], sum(pick))
+            exposure <- min(design@exposure[lastitem+1L], sum(pick))
         }
     }
     if(exposure == 1L){